feat: fix all test suites & refactor code
