+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
 docs/
-src/resourses/middlewar/access.log
-rc/resourses/middlewar/error.log
+src/resourses/middleware/access.log
+rc/resourses/middleware/error.log
