+++ src/app.ts
@@ -1,11 +1,10 @@
 import { NextFunction, Request, Response } from "express";
 
-
 const express = require('express');
-const {reqAccessLog, errorHandling} = require('./resources/middlewar/logger')
 const swaggerUI = require('swagger-ui-express');
 const path = require('path');
 const YAML = require('yamljs');
+const {reqAccessLog, errorHandling} = require('./resources/middleware/logger')
 const userRouter = require('./resources/users/user.router');
 const boardRouter = require('./resources/boards/boards.router');
 const taskRouter =  require('./resources/tasks/task.router');
@@ -13,12 +12,11 @@ const taskRouter =  require('./resources/tasks/task.router');
 const app = express();
 const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
 
-
 app.use(express.json());
 
 app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
 
-app.use('/', (_err:Error, req: Request, res: Response, next: NextFunction ) => {
+app.use('/', (req: Request, res: Response, next: NextFunction ) => {
   if (req.originalUrl === '/') {
     res.send('Service is running!');
     return;
@@ -27,13 +25,23 @@ app.use('/', (_err:Error, req: Request, res: Response, next: NextFunction ) => {
 });
 
 app.use(reqAccessLog);
-app.use (errorHandling)
 
 app.use('/users', userRouter);
 app.use ('/boards', boardRouter);
 app.use ('/boards', taskRouter);
 
+app.use (errorHandling)
 
-app.use(errorHandling)
+process.on('uncaughtException', async (err:NodeJS.ErrnoException) => {
+  await errorHandling(err);
+  process.exit(1);
+});
+
+process.on('unhandledRejection', (rs, prom) => {
+  prom.catch((err) => {
+    errorHandling(err, rs);
+    process.exit(1);
+  });
+});
 
 module.exports = app;
