--- src/resources/middlewar/error.log
@@ -1,1 +1,0 @@
-
